Move FarmerMetadata to core-primitives
