+++ crates/subspace-farmer/src/rpc.rs
@@ -4,29 +4,9 @@ use jsonrpsee::types::{Error, Subscription};
 use jsonrpsee::ws_client::{WsClient, WsClientBuilder};
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
-use subspace_core_primitives::{EncodedBlockWithObjectMapping, Salt, Tag};
+use subspace_core_primitives::{EncodedBlockWithObjectMapping, FarmerMetadata, Salt, Tag};
 
 type SlotNumber = u64;
-
-/// Metadata necessary for farmer operation
-#[derive(Debug, Deserialize)]
-pub(super) struct FarmerMetadata {
-    /// Depth `K` after which a block enters the recorded history (a global constant, as opposed
-    /// to the client-dependent transaction confirmation depth `k`).
-    pub confirmation_depth_k: u32,
-    /// The size of data in one piece (in bytes).
-    pub record_size: u32,
-    /// Recorded history is encoded and plotted in segments of this size (in bytes).
-    pub recorded_history_segment_size: u32,
-    /// This constant defines the size (in bytes) of one pre-genesis object.
-    pub pre_genesis_object_size: u32,
-    /// This constant defines the number of a pre-genesis objects that will bootstrap the
-    /// history.
-    pub pre_genesis_object_count: u32,
-    /// This constant defines the seed used for deriving pre-genesis objects that will bootstrap
-    /// the history.
-    pub pre_genesis_object_seed: Vec<u8>,
-}
 
 // There are more fields in this struct, but we only care about one
 #[derive(Debug, Deserialize)]
